documentation: Implement a robust window selection strategy
Fixes issues mentionned on
https://github.com/company-mode/company-mode/issues/309 . The goal
window is not spontaneously selected to display the documentation
buffer, and in the latest release of company selecting the target window
from the doc-buffer function causes the doc buffer to stick, instead of
being buried.

The solution here is to register a special function that checks whether
we're displaying the documentation buffer, and to show it in the coq
buffer if it has a window.