documentation: One more shr fix
shr on 24.3 and 24.4 doesn't know how to deal with shr-width =
0. Setting it to most-positive-fixnum almost works, except when shr
tries to render an <hr> (Out  of memory) or a table. Thus, removes all
tables and hrs from the source (the tables are only in the indexes), and
use most-positive-fixnums for emacs <25. For emacs >= 25, shr deals with
shr-width = 0 just fine.