completion: Add support for tactic notations
This is done by querying the prover using [Print Grammar tactic]. The
output is tricky to parse; using Wysent would probably be better (the
current code relies on a mix of converting to list code and manual
parsing). Since most of the notations are already defined in the manual,
we keep a list of tactics to ignore, built at the same time as we do
capability detection.