general: Refactor the code into a collection of small features
The new define-feature macro is used to isolate bits of independent
functionality, all of which can be turned on or off. Company-coq is now
a minor mode, supports being activated and deactivated, and exposes a
cleaner customization interface.

This was also a good occasion to document all of the code's functions,
and fix all relevant documentation warnings.