completion: Split refman backends and pre-sort refman completions
Tactics, vernacs, ltac forms and scopes are now handled by separate
backends, and sorted by the Python script that processes the manual.

Sorting is a bit smarter than it used to be: it sorts tactics by
chapter, putting argumentless ones first in each chapter, but otherwise
not changing manual order.  It compares vernacs (and ltac forms) by
stripped abbrevs (identifiers and punctuation removed), then by number
of arguments inside each chapter, and uses manual order for final
disambiguation. Note that options get a special treatment, so that Set,
Unset, and Test always come in immediate succession.

An advantage of doing this at generation time is that we can sort
variants below their original vernac. We also gain about 1/10th of a
second at startup.

Amusingly, the vernacs comparison function (-lessp-static-abbrevs)
wasn't even transitive (with two refman vernacs A and C in the same
chapter, such that C came before A in the manual, and with one PG vernac
B, it would sort C < A, but B < C and A < B).