modeline: Add a spinner and fix the icon's background color
The icons now starts spinning after a few instants of the prover being
busy. This commit also tracks the selected window, to ensure that the
background that we ask from ImageMagick is always the right one (the
issue is that Emacs handles transparency by rendering the image on a
given background, and that background wouldn't be right if the modeline
was inactive. Tracing the selected window allows us to tell whether the
current window is selected, and thus which background color we need to
use.

ImageMagick in Emacs claims to support a :rotation argument, but that
didn't seem to work, and I needed all the images to fit in the same box,
so I pre-rendered them. Emacs will cache them.