general: Make refontification faster
We don't need to font-lock-ensure for the entire buffer anymore, since
the code in -fontify-buffer is only called after enabling or diosabling
settings; that is, it isn't called from the parts that require a full,
immediate refontification before copying a string, for example (these
use -fontify-with). Thus lazy refontification is enough.

This saves a lot of time when displaying a large buffer from C-w in
company completion, or when opening a file using company-coq.