code-folding: Simplify handling of bullet keymaps
Bullets use a text property to specify their keymap (this is convenient
because the property can be added and removed very simply by
font-lock). Unfortunately, PG also assigns a keymap to some of its
overlays, and that keymap takes priority. Using a local-map instead of a
keymap worked, but only for bindings that were not overridden by any of
the minor modes in use.

This patch simply removes the keymap set by PG on its spans; it only
gave access to a context menu with an option to undo or copy the span.